+++ springboot/lms/src/main/java/com/ntuc/lms/repository/FineRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ntuc.lms.model.Fine;
 
-public interface FineRepository extends JpaRepository<Fine, Long> {
+public interface FineRepository extends JpaRepository<Fine, Integer> {
     List<Fine> findByPaidFalse();
 }
